Move everything Capture-related + Transform activity to Mods.Common
