+++ OpenRA.Mods.Common/Activities/CaptureActor.cs
@@ -8,14 +8,12 @@
  */
 #endregion
 
-using OpenRA.Mods.Common.Activities;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Activities
+namespace OpenRA.Mods.Common.Activities
 {
-	class CaptureActor : Enter
+	public class CaptureActor : Enter
 	{
 		readonly Actor actor;
 		readonly Capturable capturable;
